3.1.12

+++ Plugins/XsollaSDK/Standalone/Runtime/Core/Token/Internal/TokenData.cs
@@ -4,7 +4,7 @@ namespace Xsolla.Core
 	{
 		public string accessToken;
 		public string refreshToken;
-		public int expirationTime;
+		public long expirationTime;
 		public bool isBasedOnDeviceId;
 	}
 }
